[#7 list shows] Fix navigation from tv show episode to detail screen.
